@@ -43,20 +43,20 @@
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^2.1.0",
-    "@sveltejs/kit": "^1.20.4",
-    "@sveltejs/package": "^2.0.2",
-    "autoprefixer": "^10.4.14",
-    "postcss": "^8.4.24",
-    "publint": "^0.1.12",
-    "svelte": "^4.0.0",
-    "svelte-check": "^3.4.4",
-    "tailwindcss": "^3.3.2",
-    "tslib": "^2.5.3",
-    "typescript": "^5.1.3",
-    "vite": "^4.3.9"
+    "@sveltejs/adapter-auto": "^2.1.1",
+    "@sveltejs/kit": "^1.30.3",
+    "@sveltejs/package": "^2.2.3",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.32",
+    "publint": "^0.2.6",
+    "svelte": "next",
+    "svelte-check": "^3.6.2",
+    "tailwindcss": "^3.3.6",
+    "tslib": "^2.6.2",
+    "typescript": "^5.3.3",
+    "vite": "^4.5.1"
   },
   "peerDependencies": {
-    "svelte": "^3.59.1 || ^4.0.0"
+    "svelte": "^^5.0.0"
   }
 }
