@@ -35,36 +35,53 @@
 	import { getContext, setContext, createEventDispatcher, tick } from 'svelte'
 	import { writable } from 'svelte/store'
 
-	// state of element (shown or hidden), if null this we are treated as a child
-	// transition and will get the state from our parent, coordinating with it
-	export let show: boolean | null = null
+	interface Props {
+		// state of element (shown or hidden), if null this we are treated as a child
+		// transition and will get the state from our parent, coordinating with it
+		show?: boolean | null
 
-	// apply transition when element is first rendered (i.e. animate in)
-	export let appear: boolean = false
+		// apply transition when element is first rendered (i.e. animate in)
+		appear?: boolean
 
-	// whether the element should be removed from the DOM (vs hidden)
-	export let unmount: boolean = false
+		// whether the element should be removed from the DOM (vs hidden)
+		unmount?: boolean
 
-	// classes to apply when entering (showing)
-	export let enter: string = ''
-	export let enterFrom: string = ''
-	export let enterTo: string = ''
+		// classes to apply when entering (showing)
+		enter?: string
+		enterFrom?: string
+		enterTo?: string
 
-	// classes toi apply when leaving (hiding)
-	export let leave: string | null = null
-	export let leaveFrom: string | null = null
-	export let leaveTo: string | null = null
+		// classes to apply when leaving (hiding)
+		leave?: string | null
+		leaveFrom?: string | null
+		leaveTo?: string | null
+
+		children: any
+	}
+
+	let {
+		show = null,
+		appear = false,
+		unmount = false,
+		enter = '',
+		enterFrom = '',
+		enterTo = '',
+		leave = null,
+		leaveFrom = null,
+		leaveTo = null,
+		children,
+	} = $props<Props>()
 
 	// convert class strings to arrays, for easier use with DOM elements
-	$: enterClasses = classes(enter)
-	$: enterFromClasses = classes(enterFrom)
-	$: enterToClasses = classes(enterTo)
+	let enterClasses = $derived(classes(enter))
+	let enterFromClasses = $derived(classes(enterFrom))
+	let enterToClasses = $derived(classes(enterTo))
 
 	// if leave, leaveFrom, or leaveTo are not specified then use enter values
 	// as a convenient way to avoid repeating definitions (but reverse To & From)
-	$: leaveClasses = classes(leave === null ? enter : leave)
-	$: leaveFromClasses = classes(leaveFrom === null ? enterTo : leaveFrom)
-	$: leaveToClasses = classes(leaveTo === null ? enterFrom : leaveTo)
+	let leaveClasses = $derived(classes(leave === null ? enter : leave))
+	let leaveFromClasses = $derived(classes(leaveFrom === null ? enterTo : leaveFrom))
+	let leaveToClasses = $derived(classes(leaveTo === null ? enterFrom : leaveTo))
 
 	// get parent context if we're a child
 	const parent = show === null ? getContext<Context>(key) : null
@@ -81,8 +98,17 @@
 	}
 
 	// initial state
-	let display = show && !context.appear ? 'contents' : 'none'
-	let mounted = !unmount || show === true
+	let display = $state(show && !context.appear ? 'contents' : 'none')
+	let mounted = $state(!unmount || show === true)
+
+	$inspect({
+		show, appear, unmount, enter, enterFrom, enterTo, leave, leaveFrom, leaveTo, children,
+		enterClasses, enterFromClasses, enterToClasses,
+		leaveClasses, leaveFromClasses, leaveToClasses,
+		parent, context,
+		display,
+		mounted,
+	})
 
 	// set context for children to use
 	setContext(key, context)
@@ -245,4 +271,4 @@
 	}
 </script>
 
-<div style:display use:transition={show}>{#if mounted}<slot />{/if}</div>
+<div style:display use:transition={show}>{#if mounted}{@render children()}{/if}</div>
